I've manipulated the Simple Form Fragment code, but I feel like I'm missing one more step to properly store the form's information. I don't know what it is though.
Here is the original Simple Form Tutorial I followed, but it uses the Main Activity for reference.
http://code.tutsplus.com/tutorials/android-essentials-creating-simple-user-forms--mobile-1758

I found this project that is using a Fragment instead of an Activity, and I tried applying the same concepts to my code. I can't figure out what the missing piece is though.
http://stackoverflow.com/questions/31929138/how-do-i-findviewbyid-in-android-studio-with-fragments